Moved comment dto to service.
